+++ test/storage/append_events_test.exs
@@ -106,7 +106,7 @@ defmodule EventStore.Storage.AppendEventsTest do
     assert {:error, :wrong_expected_version} = Appender.append(conn, stream_id, events)
   end
 
-  test "append to existing stream, but stream does not exist", %{conn: conn} do
+  test "append to stream that does not exist", %{conn: conn} do
     stream_uuid = UUID.uuid4()
     stream_id = 1
     events = EventFactory.create_recorded_events(1, stream_uuid)